Refactor the actions validator a little to share some stuff with data conditions
