+++ src/sentry/workflow_engine/endpoints/validators/base/action.py
@@ -1,12 +1,10 @@
 from typing import Any, Generic, TypeVar
 
-from django.forms import ValidationError
-from jsonschema import ValidationError as JsonValidationError
-from jsonschema import validate
 from rest_framework import serializers
 
 from sentry.api.serializers.rest_framework import CamelSnakeModelSerializer
 from sentry.db.models import Model
+from sentry.workflow_engine.endpoints.validators.utils import validate_json_schema
 from sentry.workflow_engine.models import Action
 from sentry.workflow_engine.registry import action_handler_registry
 from sentry.workflow_engine.types import ActionHandler
@@ -16,38 +14,18 @@ ActionData = dict[str, Any]
 ActionConfig = dict[str, Any]
 
 
-def validate_json_schema(value, schema):
-    try:
-        validate(value, schema)
-    except JsonValidationError as e:
-        raise ValidationError(str(e))
-
-    return value
-
-
 class BaseActionValidator(CamelSnakeModelSerializer[T], Generic[T]):
     data: Any = serializers.JSONField()
     config: Any = serializers.JSONField()
+    type = serializers.ChoiceField(choices=[(t.value, t.name) for t in Action.Type])
+    integration_id = serializers.IntegerField()
 
     class Meta:
         model = T
-        fields = ["config", "data", "integration_id", "type"]
 
     def _get_action_handler(self) -> ActionHandler:
-        initial_type = self.initial_data.get("type")
-        action_type = self.validate_type(initial_type)
+        action_type = self.initial_data.get("type")
         return action_handler_registry.get(action_type)
-
-    def validate_type(self, value) -> Action.Type:
-        if not value:
-            raise ValidationError("Action type is required")
-
-        try:
-            Action.Type(value)
-        except ValueError:
-            raise ValidationError(f"Invalid Action.type, {value}")
-
-        return value
 
     def validate_data(self, value) -> ActionData:
         data_schema = self._get_action_handler().data_schema
